requirements.txt
REQUIREMENTS

streamlit==1.35.0
spacy==3.7.2
scikit-learn==1.4.2
pandas==2.2.2
PyMuPDF==1.23.22
python-docx==1.1.0
numpy==1.26.4
